+++ src/components/emails/EmailsAdmin.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import ObjectLoader from "@/components/datatable/ObjectLoader.vue";
+import ObjectTable from "@/components/datatable/ObjectTable.vue";
 import { useMainStore } from "@/stores/main";
 import { useMenuStore } from "@/stores/menu";
 import TabPanel from "primevue/tabpanel";
@@ -15,28 +15,28 @@ menuStore.setTitle("Emails");
   <div class="h-full tabview-full-height">
     <TabView lazy>
       <TabPanel :header="t('Campaigns')">
-        <ObjectLoader
+        <ObjectTable
           :store="mainStore"
           :name="'emailsCampaigns'"
           :defaultColumns="['status', 'template', 'recipients', 'automation']"
         />
       </TabPanel>
       <TabPanel :header="t('Templates')">
-        <ObjectLoader
+        <ObjectTable
           :store="mainStore"
           :name="'emailsTemplates'"
           :defaultColumns="['name', 'subject', 'design']"
         />
       </TabPanel>
       <TabPanel :header="t('Designs')">
-        <ObjectLoader
+        <ObjectTable
           :store="mainStore"
           :name="'emailsDesigns'"
           :defaultColumns="['name', 'body']"
         />
       </TabPanel>
       <TabPanel :header="t('Automations')">
-        <ObjectLoader
+        <ObjectTable
           :store="mainStore"
           :name="'emailsAutomations'"
           :defaultColumns="['name', 'extension', 'is_active']"
